requirements/requirements.txt
line-bot-sdk
flask~=1.1.2
werkzeug>=0.15.4
python-dotenv~=0.14.0
pymysql~=0.10.0
sqlalchemy~=1.3.19
pandas~=1.1.0
requests~=2.24.0
